abstracts translatetext function
